redesigned stream api (#5)
