+++ parser-core/src/byte_stream.rs
@@ -1,9 +1,9 @@
-//! This module contains everything related to code streams, which are the way compiler reads the
+//! This module contains everything related to byte streams, which are the way compiler reads the
 //! source code and bytes in general.
 //!
 //! More specific information about different submodules can be found on their respective pages.
 
 pub mod core;
 pub mod extensions;
-pub mod literal_code_stream;
+pub mod literal_byte_stream;
 pub mod view;
