+++ parser-core/src/lib.rs
@@ -1,5 +1,3 @@
-mod code_stream;
-mod error_stack;
-mod location;
-mod origin;
+pub mod byte_stream;
+pub mod error_stack;
 mod testing;
