--- parser-core/src/primitives.rs
@@ -1,1 +1,0 @@
-//! TODO: write docs
